[YS-430] refactor: providers 디렉토리 분리
